Updated import_db function
Something very strange is happening with the test data.  It keeps
failing at line 644 saying that it's only getting 1 entry.  I checked
that line and it was fine.  I deleted the line in my copy of the data (a
separate copy from the github folder) and it still had the error.