GUI 1
The two files to look at are test\test_gui.py and
test\test_gui_command_new.py  the other test_gui_command.py is an older
build I kept in case this new version breaks.  I included an about
dialog in the GUI to explain how it works.  If we manage to get it
working it should be pretty self-explanatory.