GUI 8 - Issue 6 - Solution 2
In the event that you hate the previous solution I made these 2.  They
are the same as the current GUI, but the ability to change the database
has been removed and the default path is used instead.  Let me know
which one you like better and we'll move on with that one.