Issue 7 - Create Database Upgrade
Brilliant idea assigning self.db to db.  For the sake of being cautious
I assigned the filepath of the db to self.db and then each function that
connects to the db calls some variant of sqlite3.connect(self.db).  It's
all working and it's even got a directory picker now to look all
official.